refactor HostStatusBar to more general StatusBar
Issue: AAP-21457
